% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binarize.R
\name{Binarize}
\alias{Binarize}
\title{Binarize the image}
\usage{
Binarize(image, intersection.point)
}
\arguments{
\item{image}{Image that need to be binarize}

\item{intersection.point}{Cutoff value to binarize the image}
}
\value{
Binarized image
}
\description{
Binarize the image
}
