% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowser.R
\name{bowser}
\alias{bowser}
\title{Unsupervised bayesian method to estimate the number of clusters}
\usage{
bowser(image.mat, interval.size = 0.1, r = 6, vacuum.size = 30)
}
\arguments{
\item{image.mat}{Raw image which the number of clusters should be estimated}

\item{interval.size}{Interval size that used when binarize the image}

\item{r}{Width of the neighborhood when applying the median filtering}

\item{vacuum.size}{Minimum size of clusters}
}
\value{

}
\description{
Unsupervised bayesian method to estimate the number of clusters
}
