% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoff.R
\name{getCutoff}
\alias{getCutoff}
\title{Get a intersection point from two densities}
\usage{
getCutoff(mclust.object, density1, density2, interval)
}
\arguments{
\item{mclust.object}{Object obtained from Mclust().}

\item{density1}{Density object of cluster 1 obtained from density().}

\item{density2}{Density object of cluster 2 obtained from density().}

\item{interval}{Interval to approximate the intersection point.}
}
\value{
Intersection point where the density1 and density2 cross
}
\description{
Get a intersection point from two densities
}
