% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomCluster.R
\name{getRandomCluster}
\alias{getRandomCluster}
\title{Generate random cluster for simulation}
\usage{
getRandomCluster(mask.img, background.img, Ks, locations, sizes, ints)
}
\arguments{
\item{mask.img}{Maksed image}

\item{background.img}{Image to be used to generate the normal state}

\item{Ks}{Vector of number of clusters to be simulated}

\item{locations}{Location of clusters to be simulated}

\item{sizes}{Size of clusters to be simulated}

\item{ints}{How high the clusters value should be}
}
\value{
Image of clusters with random number of clusters, locations, sizees and intensity
}
\description{
Generate random cluster for simulation
}
