% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimCluster.R
\name{getSimCluster}
\alias{getSimCluster}
\title{Generate the cluster for simulation}
\usage{
getSimCluster(mask.img, background.img, K, locations, sizes, ints)
}
\arguments{
\item{mask.img}{Maksed image}

\item{background.img}{Image to be used to generate the normal state}

\item{K}{Number of clusters to be simulated}

\item{locations}{Location of clusters to be simulated}

\item{sizes}{Size of clusters to be simulated}

\item{ints}{How high the clusters value should be}
}
\value{
Image of clusters
}
\description{
Generate the cluster for simulation
}
