% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimCohort.R
\name{getSimCohort}
\alias{getSimCohort}
\title{Generate a simulation cohort}
\usage{
getSimCohort(mask.img, background.img, N, Ks, locations, sizes, ints)
}
\arguments{
\item{mask.img}{Maksed image}

\item{background.img}{Image to be used to generate the normal state}

\item{N}{Number of images to be generated to this cohort}

\item{Ks}{Vector of number of clusters to be simulated}

\item{locations}{Location of clusters to be simulated}

\item{sizes}{Size of clusters to be simulated}

\item{ints}{How high the clusters value should be}
}
\value{
Array of images and its number of clusters
}
\description{
Generate a simulation cohort
}
