% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimCohortReplicate.R
\name{getSimCohortReplicate}
\alias{getSimCohortReplicate}
\title{Generate a simulation cohort replicates}
\usage{
getSimCohortReplicate(
  mask.img,
  background.img,
  R,
  N,
  Ks,
  locations,
  sizes,
  ints,
  parallel = FALSE,
  cores = 1,
  seed = 1
)
}
\arguments{
\item{mask.img}{Maksed image}

\item{background.img}{Image to be used to generate the normal state}

\item{R}{Number of replicates to generate}

\item{N}{Number of images to be generated to this cohort}

\item{Ks}{Vector of number of clusters to be simulated}

\item{locations}{Location of clusters to be simulated}

\item{sizes}{Size of clusters to be simulated}

\item{ints}{How high the clusters value should be}

\item{parallel}{Parallel the code or not}

\item{cores}{Number of cores for parallelization}

\item{seed}{Setting a seed for reproducibility}
}
\value{

}
\description{
Generate a simulation cohort replicates
}
