% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimData.R
\name{getSimData}
\alias{getSimData}
\title{Generate the data for simulation}
\usage{
getSimData(mask.img, background.img, location, K, size, int, set.seed = 1)
}
\arguments{
\item{mask.img}{Maksed image}

\item{background.img}{Image to be used to generate the normal state}

\item{location}{Location of clusters to be simulated}

\item{K}{Number of clusters to be simulated}

\item{size}{Size of clusters to be simulated}

\item{int}{How high the cluster value should be}

\item{set.seed}{seed}
}
\value{
Image with clusters
}
\description{
Generate the data for simulation
}
