% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc1_size5_int300_sim1.R
\docType{data}
\name{loc1_size5_int300_sim1}
\alias{loc1_size5_int300_sim1}
\title{Simulation data with disease located at (26,16)}
\format{
A matrix with 64 rows and 64 columns.
}
\source{
\url{https://github.com/ryansar/lungct}
}
\usage{
loc1_size5_int300_sim1
}
\description{
A simulated dataset with disease located at (26,16), with size 5 and intensity +300 from background.
}
\keyword{datasets}
