% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc2_size6_int400_sim1.R
\docType{data}
\name{loc2_size6_int400_sim1}
\alias{loc2_size6_int400_sim1}
\title{Simulation data with disease located at (18,31)}
\format{
A matrix with 64 rows and 64 columns.
}
\source{
\url{https://github.com/ryansar/lungct}
}
\usage{
loc2_size6_int400_sim1
}
\description{
A simulated dataset with disease located at (18,31), with size 6 and intensity +400 from background.
}
\keyword{datasets}
