% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc3_size4_int200_sim1.R
\docType{data}
\name{loc3_size4_int200_sim1}
\alias{loc3_size4_int200_sim1}
\title{Simulation data with disease located at (31,46)}
\format{
A matrix with 64 rows and 64 columns.
}
\source{
\url{https://github.com/ryansar/lungct}
}
\usage{
loc3_size4_int200_sim1
}
\description{
A simulated dataset with disease located at (31,46), with size 4 and intensity +200 from background.
}
\keyword{datasets}
