% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc4_size3_int300_sim1.R
\docType{data}
\name{loc4_size3_int300_sim1}
\alias{loc4_size3_int300_sim1}
\title{Simulation data with disease located at (40,21)}
\format{
A matrix with 64 rows and 64 columns.
}
\source{
\url{https://github.com/ryansar/lungct}
}
\usage{
loc4_size3_int300_sim1
}
\description{
A simulated dataset with disease located at (40,21), with size 3 and intensity +300 from background.
}
\keyword{datasets}
