% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc5_size2_int200_sim1.R
\docType{data}
\name{loc5_size2_int200_sim1}
\alias{loc5_size2_int200_sim1}
\title{Simulation data with disease located at (40,57)}
\format{
A matrix with 64 rows and 64 columns.
}
\source{
\url{https://github.com/ryansar/lungct}
}
\usage{
loc5_size2_int200_sim1
}
\description{
A simulated dataset with disease located at (40,57), with size 2 and intensity +200 from background.
}
\keyword{datasets}
