% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_scan.R
\name{reduce_scan}
\alias{reduce_scan}
\title{Function used to reduce images}
\usage{
reduce_scan(img, mask)
}
\arguments{
\item{img}{Name of the file to read the image from}

\item{mask}{Name of the file to read the mask from}
}
\value{
Lists of reduced image
}
\description{
Function used to reduce images
}
