% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vacuum.R
\name{Vacuum}
\alias{Vacuum}
\title{Remove the noise}
\usage{
Vacuum(image, size = 30)
}
\arguments{
\item{image}{Matrix (or array) with connected components}

\item{size}{Minimum number of pixels (or voxels) to be considered as cluster}
}
\value{
Matrix (or array) with noise removed
}
\description{
Remove the noise
}
