% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimData.R
\name{getSimData}
\alias{getSimData}
\title{Generate the data for simulation}
\usage{
getSimData(mask.img, background.img, location, size, int)
}
\arguments{
\item{mask.img}{Maksed image}

\item{background.img}{Image to be used to generate the normal state}

\item{location}{Location of cluster to be simulated}

\item{size}{Size of cluster to be simulated}

\item{int}{How high the cluster value should be}
}
\value{
Image with cluster
}
\description{
Generate the data for simulation
}
